requirements.txt
diff-match-patch==20121119
dj-database-url==0.4.1
Django==1.9.6
django-cors-headers==1.1.0
django-excel==0.0.4
django-extensions==1.6.6
django-filter==0.13.0
django-import-export==0.4.5
djangorestframework==3.3.3
future==0.15.2
gunicorn==19.4.5
jdcal==1.2
Markdown==2.6.6
odfpy==0.9.6
openpyxl==2.2.2
psycopg2==2.6.1
pyexcel==0.2.1
pyexcel-io==0.1.0
pyexcel-ods==0.1.1
pyexcel-webio==0.0.6
pyexcel-xls==0.1.0
pyexcel-xlsx==0.1.0
six==1.10.0
tablib==0.11.2
texttable==0.8.4
whitenoise==3.0
xlrd==0.9.4
xlwt-future==0.8.0
